refactor: State 패키지 이동
