+++ src/main/java/chess/game/TurnStrategy.java
@@ -1,6 +1,6 @@
 package chess.game;
 
-import chess.game.state.RunningState;
+import chess.game.state.running.RunningState;
 
 public interface TurnStrategy {
     RunningState create();
