+++ src/main/java/chess/game/state/end/NoneWinState.java
@@ -1,25 +1,25 @@
-package chess.game.state;
+package chess.game.state.end;
 
 import chess.domain.Team;
 import chess.dto.SquareResponse;
 import java.util.List;
 import java.util.function.Supplier;
 
-public class BlackWinState extends EndState {
-    public static final EndState STATE = new BlackWinState();
+public class NoneWinState extends EndState {
+    public static final EndState STATE = new NoneWinState();
 
     @Override
     public Team getWinner() {
-        return Team.BLACK;
+        throw new IllegalStateException(END_STATE_EXCEPTION_MESSAGE);
     }
 
     @Override
     public List<SquareResponse> getBoard(Supplier<List<SquareResponse>> supplier) {
-        return supplier.get();
+        throw new IllegalStateException(END_STATE_EXCEPTION_MESSAGE);
     }
 
     @Override
     public boolean hasWinner() {
-        return true;
+        return false;
     }
 }
