+++ src/main/java/chess/game/state/end/WhiteWinState.java
@@ -1,25 +1,25 @@
-package chess.game.state;
+package chess.game.state.end;
 
 import chess.domain.Team;
 import chess.dto.SquareResponse;
 import java.util.List;
 import java.util.function.Supplier;
 
-public class NoneWinState extends EndState {
-    public static final EndState STATE = new NoneWinState();
+public class WhiteWinState extends EndState {
+    public static final EndState STATE = new WhiteWinState();
 
     @Override
     public Team getWinner() {
-        throw new IllegalStateException(END_STATE_EXCEPTION_MESSAGE);
+        return Team.WHITE;
     }
 
     @Override
     public List<SquareResponse> getBoard(Supplier<List<SquareResponse>> supplier) {
-        throw new IllegalStateException(END_STATE_EXCEPTION_MESSAGE);
+        return supplier.get();
     }
 
     @Override
     public boolean hasWinner() {
-        return false;
+        return true;
     }
 }
