+++ src/test/java/chess/game/MockTurnStrategy.java
@@ -1,6 +1,6 @@
 package chess.game;
 
-import chess.game.state.RunningState;
+import chess.game.state.running.RunningState;
 
 public class MockTurnStrategy implements TurnStrategy {
